STRWEB-73 remove babel-plugin-lodash, lodash-webpack-plugin (#89)
Remove `babel-plugin-lodash` and `lodash-webpack-plugin`. They only
offered a 0.5% reduction in bundle-size, which hardly seems worth
the effort to maintain them.

WRT `babel-plugin-lodash`:

This unmaintained plugin now spews thousands of lines of alerts during
install due to one of the webpack functions it leverages being saddled
with a deprecation warning. The fix is utterly trivial, but the
maintainers have been silent for weeks, not providing any faith that
this will ever be addressed.

It was nice to have a small lodash footprint in our bundle, but it felt
like a microoptimization given our bundlesize, and something that was
not worth the benefit given the poor DX.

This effectively undoes STRWEB-20/PR #28

WRT `lodash-webpack-plugin`:

This plugin also appears unmaintained. It offers the ability to include
select feature sets in lodash, but our config includes every optional
piece; IOW, not really doing anything for us.

Refs STRWEB-73