STRWEB-102 forward stripes-config values to service worker (#132)
Use a virtual module to forward some stripes-config values to the
service worker when it is bundled. These values must be compiled in
because global variables in a service worker are volatile and are thus
reset to their default (null) values on each sleep/wake cycle. Thus, the
only way to make these values permanently available is to compile them
in.

h/t @mkuklis and @JohnC-80 who did the heavy lifting here, helping
to think through potential solutions and figuring out how to actually
implement them in our highly customized build.

Refs STRWEB-102, STCOR-759

---------

Co-authored-by: Michal Kuklis <michal.kuklis@gmail.com>